==> 04_ML models.ipynb <==
import os
import pandas as pd
from joblib import dump
from sklearn.model_selection import ShuffleSplit
from sklearn.model_selection import RandomizedSearchCV
from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import r2_score

parent_dir = os.path.dirname(os.getcwd())

# Train data preparation

# Initialize directories
models_dir = os.path.join(parent_dir, "Trained Models")
os.makedirs(models_dir, exist_ok=True)
data_dir = os.path.join(parent_dir, "Data")

# Load train data set
train_df = pd.read_csv(os.path.join(data_dir, "train.csv"))
train_df.head()

# Drop the first three columns (id, k, and arc)
train = train_df.drop(columns=["id", "k", "arc"])

# Split data into 75% train and 25% test
rs = ShuffleSplit(n_splits=1, test_size=.25, random_state=0)
for train_index, test_index in rs.split(train):
    train_set = train.loc[train_index]
    test_set = train.loc[test_index]
print(f"Sizes: data set {train.shape}, train set {train_set.shape}, test set {test_set.shape}")

# Determine the predictors and target
X_train = train_set.drop(columns=["F_r"])
y_train = train_set["F_r"]
X_test = test_set.drop(columns=["F_r"])
y_test = test_set["F_r"]

# Decision tree model

# Use random search to find the best hyperparameters for the decision tree
param_dist = {"max_depth": range(1, 21)}
tree = DecisionTreeRegressor()
tree_cv = RandomizedSearchCV(tree, param_dist, cv=5, n_iter=20, random_state=0)
tree_cv.fit(X_train, y_train)
print(f"Best hyperparameters: {tree_cv.best_params_['max_depth']}")

# Construct the decision tree model
dt = DecisionTreeRegressor(random_state=0, max_depth=tree_cv.best_params_['max_depth'])
dt.fit(X_train, y_train)
y_pred = dt.predict(X_test)
r2 = r2_score(y_test, y_pred)
print(f"Decision Tree R2: {r2}")

# Save the decision tree model
model_path = os.path.join(models_dir, "dt.joblib")
dump(dt, model_path)

# Random forest model

# Use random search to find the best hyperparameters for the random forest
param_dist = {"max_depth": range(1, 21)}
rf = RandomForestRegressor()
rf_cv = RandomizedSearchCV(rf, param_dist, cv=5, n_iter=20, random_state=0)
rf_cv.fit(X_train, y_train)
print(f"Best hyperparameters: {rf_cv.best_params_['max_depth']}")

# Construct the random forest model
rf = RandomForestRegressor(random_state=0, max_depth=rf_cv.best_params_['max_depth'])
rf.fit(X_train, y_train)
y_pred = rf.predict(X_test)
r2 = r2_score(y_test, y_pred)
print(f"Random Forest R2: {r2}")

# Save the random forest model
model_path = os.path.join(models_dir, "rf.joblib")
dump(rf, model_path)
