==> 05_Deploy ML models.ipynb <==
import os
import pandas as pd
from joblib import load

# Initialize directories
parent_dir = os.path.dirname(os.getcwd())
models_dir = os.path.join(parent_dir, "Trained Models")
data_dir = os.path.join(parent_dir, "Data")
predictions_dir = os.path.join(parent_dir, "Predictions")
os.makedirs(predictions_dir, exist_ok=True)

# Import ML models

# Decision tree model
dt = load(os.path.join(models_dir, "dt.joblib"))
# Random forest model
rf = load(os.path.join(models_dir, "rf.joblib"))

# Predict using train data

## Train data preparation

# Load train data set
train = pd.read_csv(os.path.join(data_dir, "train.csv"))
train.head()

# Drop the first three columns (id, k, and arc) as well as the target variable (F_r)
X_train = train.drop(columns=["id", "k", "arc", "F_r"])

## Predict the target variable (reduction percentage in max flow)

### Decision tree model


# Initialize directories
train_dir = os.path.join(predictions_dir, "dt", "Train")
os.makedirs(train_dir, exist_ok=True)

# Predictions
train_dt = train.copy()
train_dt["dt"] = dt.predict(X_train)

# Save the predictions for each network
groups = train_dt.groupby(["id", "k"])
for (id, k), group in groups:
    group.to_csv(os.path.join(train_dir, f"network_{id}_{k}.csv"), index=False)

### Random forest model

# Initialize directories
train_dir = os.path.join(predictions_dir, "rf", "Train")
os.makedirs(train_dir, exist_ok=True)

# Predictions
train_rf = train.copy()
train_rf["rf"] = rf.predict(X_train)

# Save the predictions for each network
groups = train_rf.groupby(["id", "k"])
for (id, k), group in groups:
    group.to_csv(os.path.join(train_dir, f"network_{id}_{k}.csv"), index=False)

# Predict using test data

## Test data preparation

# Load test data set
test = pd.read_csv(os.path.join(data_dir, "test.csv"))
test.head()

# Drop the first three columns (id, k, and arc)
X_test = test.drop(columns=["id", "k", "arc"])

# Predict the target variable (reduction percentage in max flow)

### Decision tree model


# Initialize directories
test_dir = os.path.join(predictions_dir, "dt", "Test")
os.makedirs(test_dir, exist_ok=True)

# Predictions
test_dt = test.copy()
test_dt["dt"] = dt.predict(X_test)

# Save the predictions for each network
groups = test_dt.groupby(["id", "k"])
for (id, k), group in groups:
    group.to_csv(os.path.join(test_dir, f"network_{id}_{k}.csv"), index=False)

### Random forest model

### Decision tree model


# Initialize directories
train_dir = os.path.join(predictions_dir, "dt", "Train")
os.makedirs(train_dir, exist_ok=True)

# Predictions
train_dt = train.copy()
train_dt["dt"] = dt.predict(X_train)

# Save the predictions for each network
groups = train_dt.groupby(["id", "k"])
for (id, k), group in groups:
    group.to_csv(os.path.join(train_dir, f"network_{id}_{k}.csv"), index=False)

### Random forest model

# Initialize directories
train_dir = os.path.join(predictions_dir, "rf", "Train")
os.makedirs(train_dir, exist_ok=True)

# Predictions
train_rf = train.copy()
train_rf["rf"] = rf.predict(X_train)

# Save the predictions for each network
groups = train_rf.groupby(["id", "k"])
for (id, k), group in groups:
    group.to_csv(os.path.join(train_dir, f"network_{id}_{k}.csv"), index=False)

# Initialize directories
test_dir = os.path.join(predictions_dir, "rf", "Test")
os.makedirs(test_dir, exist_ok=True)

# Predictions
test_rf = test.copy()
test_rf["rf"] = rf.predict(X_test)

# Save the predictions for each network
groups = test_rf.groupby(["id", "k"])
for (id, k), group in groups:
    group.to_csv(os.path.join(test_dir, f"network_{id}_{k}.csv"), index=False)
