==> 06_Solve using ML models.ipynb <==
import os
import pandas as pd
import networkx as nx
from tqdm import tqdm

from NetworkGen import *
from GurobiSolver import *
from GreedyInterdictAlg import *

parent_dir = os.path.dirname(os.getcwd())

# Test data sets

## Decision tree model

# Initialize directories
data = "Test"
ml_alg = "dt"
network_dir = os.path.join(parent_dir, "Networks", data)
predictions_dir = os.path.join(parent_dir, "Predictions", ml_alg, data)
models_dir = os.path.join(parent_dir, "Models-ml", ml_alg, data)
os.makedirs(models_dir, exist_ok=True)

### Arc interdiction costs are set to 1

unit_cost = True
tag = "U"

# Interdiction budget
small_uc_B = [i for i in range(1, 9+1)]
large_uc_B = [1, 10, 50, 100, 500, 1000, 2000, 2500, 3000]

# Solve instances
for f in tqdm(os.listdir(network_dir)):
    name, ext = f.split('.')
    G = load_network(f"{network_dir}/{name}")
    if name.split('_')[1] in map(str, [6, 7, 8]):
        B_vals = large_uc_B
    else:
        B_vals = small_uc_B
    for B in B_vals:
        # Make directories
        model_dir = os.path.join(models_dir, f"Model_{B}{tag}")
        for format in ["mps", "sol", "lp", "json"]:
            os.makedirs(f"{model_dir}/{format}", exist_ok=True)
        # Find the interdicted arcs
        interdicted_arcs = greedy_alg(G, f"{predictions_dir}/{name}", B, unit_cost, ml_alg)
        # Solve the model
        m = find_max_flow(G, name, model_dir, B, interdicted_arcs)

### Arc interdiction costs are greater that or equal to 1

unit_cost = False
tag = ""

# Interdiction budget
small_B = [5, 10, 20, 30, 40, 80, 120, 150, 200]
large_B = [5, 100, 200, 500, 1000, 2000, 3000, 4000, 5000, 7500, 10000, 15000, 20000]

# Solve instances
for f in tqdm(os.listdir(network_dir)):
    name, ext = f.split('.')
    G = load_network(f"{network_dir}/{name}")
    if name.split('_')[1] in map(str, [6, 7, 8]):
        B_vals = large_B
    else:
        B_vals = small_B
    for B in B_vals:
        # Make directories
        model_dir = os.path.join(models_dir, f"Model_{B}{tag}")
        for format in ["mps", "sol", "lp", "json"]:
            os.makedirs(f"{model_dir}/{format}", exist_ok=True)
        # Find the interdicted arcs
        interdicted_arcs = greedy_alg(G, f"{predictions_dir}/{name}", B, unit_cost, ml_alg)
        # Solve the model
        m = find_max_flow(G, name, model_dir, B, interdicted_arcs)

## Random forest model

# Initialize directories
data = "Test"
ml_alg = "rf"
network_dir = os.path.join(parent_dir, "Networks", data)
predictions_dir = os.path.join(parent_dir, "Predictions", ml_alg, data)
models_dir = os.path.join(parent_dir, "Models-ml", ml_alg, data)
os.makedirs(models_dir, exist_ok=True)

### Arc interdiction costs are set to 1

unit_cost = True
tag = "U"

# Interdiction budget
small_uc_B = [i for i in range(1, 9+1)]
large_uc_B = [1, 10, 50, 100, 500, 1000, 2000, 2500, 3000]

# Solve instances
for f in tqdm(os.listdir(network_dir)):
    name, ext = f.split('.')
    G = load_network(f"{network_dir}/{name}")
    if name.split('_')[1] in map(str, [6, 7, 8]):
        B_vals = large_uc_B
    else:
        B_vals = small_uc_B
    for B in B_vals:
        # Make directories
        model_dir = os.path.join(models_dir, f"Model_{B}{tag}")
        for format in ["mps", "sol", "lp", "json"]:
            os.makedirs(f"{model_dir}/{format}", exist_ok=True)
        # Find the interdicted arcs
        interdicted_arcs = greedy_alg(G, f"{predictions_dir}/{name}", B, unit_cost, ml_alg)
        # Solve the model
        m = find_max_flow(G, name, model_dir, B, interdicted_arcs)

### Arc interdiction costs are greater that or equal to 1

unit_cost = False
tag = ""

# Interdiction budget
small_B = [5, 10, 20, 30, 40, 80, 120, 150, 200]
large_B = [5, 100, 200, 500, 1000, 2000, 3000, 4000, 5000, 7500, 10000, 15000, 20000]

# Solve instances
for f in tqdm(os.listdir(network_dir)):
    name, ext = f.split('.')
    G = load_network(f"{network_dir}/{name}")
    if name.split('_')[1] in map(str, [6, 7, 8]):
        B_vals = large_B
    else:
        B_vals = small_B
    for B in B_vals:
        # Make directories
        model_dir = os.path.join(models_dir, f"Model_{B}{tag}")
        for format in ["mps", "sol", "lp", "json"]:
            os.makedirs(f"{model_dir}/{format}", exist_ok=True)
        # Find the interdicted arcs
        interdicted_arcs = greedy_alg(G, f"{predictions_dir}/{name}", B, unit_cost, ml_alg)
        # Solve the model
        m = find_max_flow(G, name, model_dir, B, interdicted_arcs)
