==> 07_Obtain results.ipynb <==
import os
import pandas as pd
from tqdm import tqdm

from GurobiSolver import *

parent_dir = os.path.dirname(os.getcwd())
os.makedirs(parent_dir + '/Results', exist_ok=True)

# Train data sets

## Networks data

# Load the networks details
tr_net_df = pd.read_csv("Results/TrainNetworksDetails.csv")
tr_net_df.head()

## Decision tree model

# # Initialize directories
# data = "Train"
# ml_alg = "dt"
# network_dir = os.path.join(parent_dir, "Networks", data)
# predictions_dir = os.path.join(parent_dir, "Predictions", ml_alg, data)
# models_dir = os.path.join(parent_dir, "Models-ml", ml_alg, data)
# os.makedirs(models_dir, exist_ok=True)

## Random forest model

# # Initialize directories
# data = "Train"
# ml_alg = "rf"
# network_dir = os.path.join(parent_dir, "Networks", data)
# predictions_dir = os.path.join(parent_dir, "Predictions", ml_alg, data)
# models_dir = os.path.join(parent_dir, "Models-ml", ml_alg, data)
# os.makedirs(models_dir, exist_ok=True)

# Test data sets

## Small networks

# Load the small networks details
st_net_df = pd.read_csv("Results/SmallTestNetworksDetails.csv")
st_net_df.head()

# Initialize directories
data = "Test"
opt_models_dir = os.path.join(parent_dir, "Models-gp", data)
dt_models_dir = os.path.join(parent_dir, "Models-ml", "dt", data)
rf_models_dir = os.path.join(parent_dir, "Models-ml", "rf", data)

### Arc interdiction costs are set to 1

tag = "U"

# Interdiction budget
small_uc_B = [i for i in range(1, 9+1)]

# Results for the small networks with unit costs
cols = ["id", "k", f"B_{tag}", f"Opt_{tag}", f"ObjVal_dt_{tag}", f"ObjVal_rf_{tag}"]
vals = []
for id, k in tqdm(zip(st_net_df["id"], st_net_df["k"])):
    name = f"network_{id}_{k}"
    for B in small_uc_B:
        # Set models directory
        # Optimal model
        opt_model_dir = os.path.join(opt_models_dir, f"Model_{B}{tag}")
        opt_obj = load_json(name, opt_model_dir)["ObjVal"]
        # Decision tree model
        dt_model_dir = os.path.join(dt_models_dir, f"Model_{B}{tag}")
        dt_obj = load_json(name, dt_model_dir)["ObjVal"]
        # Random forest model
        rf_model_dir = os.path.join(rf_models_dir, f"Model_{B}{tag}")
        rf_obj = load_json(name, rf_model_dir)["ObjVal"]
        # Append results
        vals.append([id, k, B, opt_obj, dt_obj, rf_obj])
u_small_df = pd.DataFrame(vals, columns=cols)
u_small_df = u_small_df.sort_values(by=["id", f"B_{tag}", "k"])
u_small_df[f"Gap_dt_{tag}"] = (u_small_df[f"ObjVal_dt_{tag}"] - u_small_df[f"Opt_{tag}"]) / u_small_df[f"Opt_{tag}"] * 100
u_small_df[f"Gap_rf_{tag}"] = (u_small_df[f"ObjVal_rf_{tag}"] - u_small_df[f"Opt_{tag}"]) / u_small_df[f"Opt_{tag}"] * 100
u_small_df.round(2).to_csv(f"Results/SmallTestResults_{tag}.csv", index=False)

### Arc interdiction costs are greater that or equal to 1

tag = ""

# Interdiction budget
small_B = [5, 10, 20, 30, 40, 80, 120, 150, 200]

# Results for the small networks with costs greater than or equal to 1
cols = ["id", "k", f"B_{tag}", f"Opt_{tag}", f"ObjVal_dt_{tag}", f"ObjVal_rf_{tag}"]
vals = []
for id, k in tqdm(zip(st_net_df["id"], st_net_df["k"])):
    name = f"network_{id}_{k}"
    for B in small_B:
        # Set models directory
        # Optimal model
        opt_model_dir = os.path.join(opt_models_dir, f"Model_{B}{tag}")
        opt_obj = load_json(name, opt_model_dir)["ObjVal"]
        # Decision tree model
        dt_model_dir = os.path.join(dt_models_dir, f"Model_{B}{tag}")
        dt_obj = load_json(name, dt_model_dir)["ObjVal"]
        # Random forest model
        rf_model_dir = os.path.join(rf_models_dir, f"Model_{B}{tag}")
        rf_obj = load_json(name, rf_model_dir)["ObjVal"]
        # Append results
        vals.append([id, k, B, opt_obj, dt_obj, rf_obj])
small_df = pd.DataFrame(vals, columns=cols)
small_df = small_df.sort_values(by=["id", f"B_{tag}", "k"])
small_df[f"Gap_dt_{tag}"] = (small_df[f"ObjVal_dt_{tag}"] - small_df[f"Opt_{tag}"]) / small_df[f"Opt_{tag}"] * 100
small_df[f"Gap_rf_{tag}"] = (small_df[f"ObjVal_rf_{tag}"] - small_df[f"Opt_{tag}"]) / small_df[f"Opt_{tag}"] * 100
small_df.round(2).to_csv(f"Results/SmallTestResults_{tag}.csv", index=False)

## Large networks

# Load the large networks details
# At each level of the network, the number of nodes, flow capacity range, and interdiction cost range are specified in a respective order.
lt_net_df = pd.read_csv("Results/LargeTestNetworksDetails.csv")
lt_net_df.head()

# Initialize directories
data = "Test"
opt_models_dir = os.path.join(parent_dir, "Models-gp", data)
dt_models_dir = os.path.join(parent_dir, "Models-ml", "dt", data)
rf_models_dir = os.path.join(parent_dir, "Models-ml", "rf", data)

### Arc interdiction costs are set to 1

unit_cost = True
tag = "U"

# Interdiction budget
large_uc_B = [1, 10, 50, 100, 500, 1000, 2000, 2500, 3000]

# Results for the large networks with unit costs
cols = ["id", "k", f"B_{tag}", f"Opt_{tag}", f"Opt_T_{tag}", f"ObjVal_dt_{tag}", f"dt_T_{tag}", f"ObjVal_rf_{tag}", f"rf_T_{tag}"]
vals = []
for id, k in tqdm(zip(lt_net_df["id"], lt_net_df["k"])):
    name = f"network_{id}_{k}"
    for B in large_uc_B:
        # Set models directory
        # Optimal model
        opt_model_dir = os.path.join(opt_models_dir, f"Model_{B}{tag}")
        opt_json = load_json(name, opt_model_dir)
        opt_obj = opt_json["ObjVal"]
        opt_time = opt_json["Runtime"]
        # Decision tree model
        dt_model_dir = os.path.join(dt_models_dir, f"Model_{B}{tag}")
        dt_json = load_json(name, dt_model_dir)
        dt_obj = dt_json["ObjVal"]
        dt_time = dt_json["Runtime"]
        # Random forest model
        rf_model_dir = os.path.join(rf_models_dir, f"Model_{B}{tag}")
        rf_json = load_json(name, rf_model_dir)
        rf_obj = rf_json["ObjVal"]
        rf_time = rf_json["Runtime"]
        # Append results
        vals.append([id, k, B, opt_obj, opt_time, dt_obj, dt_time, rf_obj, rf_time])
u_large_df = pd.DataFrame(vals, columns=cols)
u_large_df = u_large_df.sort_values(by=["id", f"B_{tag}", "k"])
u_large_df[f"Opt_{tag}"] = u_large_df[f"Opt_{tag}"].replace(0, None) # Replace zero values with a NaN to avoid division by zero
u_large_df[f"Gap_dt_{tag}"] = (u_large_df[f"ObjVal_dt_{tag}"] - u_large_df[f"Opt_{tag}"]) / u_large_df[f"Opt_{tag}"] * 100
u_large_df[f"Gap_rf_{tag}"] = (u_large_df[f"ObjVal_rf_{tag}"] - u_large_df[f"Opt_{tag}"]) / u_large_df[f"Opt_{tag}"] * 100
u_large_df.round(2).to_csv(f"Results/LargeTestResults_{tag}.csv", index=False)

### Arc interdiction costs are greater that or equal to 1

unit_cost = False
tag = ""

# Interdiction budget
large_B = [5, 100, 200, 500, 1000, 2000, 3000, 4000, 5000, 7500, 10000, 15000, 20000]

# Results for the large networks with costs greater than or equal to 1
cols = ["id", "k", f"B_{tag}", f"Opt_{tag}", f"Opt_T_{tag}", f"ObjVal_dt_{tag}", f"dt_T_{tag}", f"ObjVal_rf_{tag}", f"rf_T_{tag}"]
vals = []
for id, k in tqdm(zip(lt_net_df["id"], lt_net_df["k"])):
    name = f"network_{id}_{k}"
    for B in large_B:
        # Set models directory
        # Optimal model
        opt_model_dir = os.path.join(opt_models_dir, f"Model_{B}{tag}")
        opt_json = load_json(name, opt_model_dir)
        opt_obj = opt_json["ObjVal"]
        opt_time = opt_json["Runtime"]
        # Decision tree model
        dt_model_dir = os.path.join(dt_models_dir, f"Model_{B}{tag}")
        dt_json = load_json(name, dt_model_dir)
        dt_obj = dt_json["ObjVal"]
        dt_time = dt_json["Runtime"]
        # Random forest model
        rf_model_dir = os.path.join(rf_models_dir, f"Model_{B}{tag}")
        rf_json = load_json(name, rf_model_dir)
        rf_obj = rf_json["ObjVal"]
        rf_time = rf_json["Runtime"]
        # Append results
        vals.append([id, k, B, opt_obj, opt_time, dt_obj, dt_time, rf_obj, rf_time])
large_df = pd.DataFrame(vals, columns=cols)
large_df = large_df.sort_values(by=["id", f"B_{tag}", "k"])
large_df[f"Opt_{tag}"] = large_df[f"Opt_{tag}"].replace(0, None) # Replace zero values with a NaN to avoid division by zero
large_df[f"Gap_dt_{tag}"] = (large_df[f"ObjVal_dt_{tag}"] - large_df[f"Opt_{tag}"]) / large_df[f"Opt_{tag}"] * 100
large_df[f"Gap_rf_{tag}"] = (large_df[f"ObjVal_rf_{tag}"] - large_df[f"Opt_{tag}"]) / large_df[f"Opt_{tag}"] * 100
large_df.round(2).to_csv(f"Results/LargeTestResults_{tag}.csv", index=False)
