==> 08_Analysis.ipynb <==
import numpy as np
import pandas as pd

# Small instances

## Arc interdiction costs are equal to 1

# Import results for small instances with unit interdiction costs
u_small_df = pd.read_csv('Results/SmallTestResults_U.csv').round(2)
# Ignore instances where gap is too large (i.e. optimal objective value is zero and ml model predicts a non-zero value)
u_small_df.replace(np.inf, np.nan, inplace=True)
u_small_df.head()

# Group results by instance and budget, then calculate average values
u_small_summary = u_small_df.groupby(["id", "B_U"]).mean().round(2)
# Some cleaning
u_small_summary.reset_index(inplace=True)
u_small_summary.replace("NaN", 0, inplace=True)
u_small_summary.drop(columns=["k", "Opt_U", "ObjVal_dt_U", "ObjVal_rf_U"], inplace=True)
u_small_summary.head()

# Check if dt and rf find the optimal solution
u_small_df['dt_optimal'] = u_small_df['ObjVal_dt_U'] == u_small_df['Opt_U']
u_small_df['rf_optimal'] = u_small_df['ObjVal_rf_U'] == u_small_df['Opt_U']

# Add a column to the summary table to indicate if the optimal solution was found
u_small_summary['dt_optimal'] = u_small_df.groupby(["id", "B_U"])['dt_optimal'].sum().values.round(2)
u_small_summary['rf_optimal'] = u_small_df.groupby(["id", "B_U"])['rf_optimal'].sum().values.round(2)
u_small_summary.replace(np.nan, "-", inplace=True)
u_small_summary["dt_U"] = u_small_summary['Gap_dt_U'].astype(str) + '/' + u_small_summary['dt_optimal'].astype(str)
u_small_summary["rf_U"] = u_small_summary['Gap_rf_U'].astype(str) + '/' + u_small_summary['rf_optimal'].astype(str)
u_small_summary.head()

## Arc interdiction costs are greater that or equal to 1

# Import results for small instances with unit interdiction costs
small_df = pd.read_csv('Results/SmallTestResults_.csv').round(2)
# Ignore instances where gap is too large (i.e. optimal objective value is zero and ml model predicts a non-zero value)
small_df.replace(np.inf, np.nan, inplace=True)
small_df.head()

# Group results by instance and budget, then calculate average values
small_summary = small_df.groupby(["id", "B_"]).mean().round(2)
# Some cleaning
small_summary.reset_index(inplace=True)
small_summary.replace("NaN", 0, inplace=True)
small_summary.drop(columns=["k", "Opt_", "ObjVal_dt_", "ObjVal_rf_"], inplace=True)
small_summary.head()

# Check if dt and rf find the optimal solution
small_df['dt_optimal'] = small_df['ObjVal_dt_'] == small_df['Opt_']
small_df['rf_optimal'] = small_df['ObjVal_rf_'] == small_df['Opt_']

# Add a column to the summary table to indicate if the optimal solution was found
small_summary['dt_optimal'] = small_df.groupby(["id", "B_"])['dt_optimal'].sum().values.round(2)
small_summary['rf_optimal'] = small_df.groupby(["id", "B_"])['rf_optimal'].sum().values.round(2)
small_summary.replace(np.nan, "-", inplace=True)
small_summary["dt_"] = small_summary['Gap_dt_'].astype(str) + '/' + small_summary['dt_optimal'].astype(str)
small_summary["rf_"] = small_summary['Gap_rf_'].astype(str) + '/' + small_summary['rf_optimal'].astype(str)
small_summary.head()

# Save the summary table
small_summary_df = pd.concat([u_small_summary[["id", "B_U", "dt_U", "rf_U"]],small_summary[["B_", "dt_", "rf_"]]], axis = 1)
small_summary_df.round(2).to_csv("Results/SmallSummary.csv", index=False)
small_summary_df.head()

# Large instances

## Arc interdiction costs are equal to 1

# Import results for large instances with unit interdiction costs
u_large_df = pd.read_csv('Results/LargeTestResults_U.csv').round(2)
# Ignore instances where gap is too large (i.e. optimal objective value is zero and ml model predicts a non-zero value)
u_large_df.replace(np.inf, np.nan, inplace=True)
u_large_df.head()

# Group results by instance and budget, then calculate average values
u_large_summary = u_large_df.groupby(["id", "B_U"]).mean().round(2)
# Some cleaning
u_large_summary.reset_index(inplace=True)
u_large_summary.replace("NaN", 0, inplace=True)
u_large_summary.drop(columns=["k", "Opt_U", "ObjVal_dt_U", "ObjVal_rf_U"], inplace=True)
u_large_summary.replace(np.nan, "-", inplace=True)
# Save the summary table
u_large_summary.round(2).to_csv("Results/LargeSummary_U.csv", index=False)
u_large_summary.head()

## Arc interdiction costs are greater that or equal to 1

# Import results for large instances with unit interdiction costs
large_df = pd.read_csv('Results/LargeTestResults_.csv').round(2)
# Ignore instances where gap is too large (i.e. optimal objective value is zero and ml model predicts a non-zero value)
large_df.replace(np.inf, np.nan, inplace=True)
large_df.head()

# Group results by instance and budget, then calculate average values
large_summary = large_df.groupby(["id", "B_"]).mean().round(2)
# Some cleaning
large_summary.reset_index(inplace=True)
large_summary.replace("NaN", 0, inplace=True)
large_summary.drop(columns=["k", "Opt_", "ObjVal_dt_", "ObjVal_rf_"], inplace=True)
large_summary.replace(np.nan, "-", inplace=True)
# Save the summary table
large_summary.round(2).to_csv("Results/LargeSummary_.csv", index=False)
large_summary.head()

# Save the summary table
large_summary_df = pd.concat([u_large_summary[["id", "B_U", "Opt_T_U", "dt_T_U", "rf_T_U", "Gap_dt_U", "Gap_rf_U"]],large_summary[large_summary["B_"].isin([5, 100, 500, 1000, 3000, 5000, 7500, 10000, 20000])].reset_index()[["B_", "Opt_T_", "dt_T_", "rf_T_", "Gap_dt_", "Gap_rf_"]]], axis = 1)
large_summary_df.round(2).to_csv("Results/LargeSummary.csv", index=False)
large_summary_df.head()
